+++ lib/router/router.dart
@@ -1,0 +1,3 @@
+export 'app_hero_tags.dart';
+export 'app_router.dart';
+export 'app_routes.dart';
