+++ lib/ui/pages/auth/widgets/sign_in_form.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:over_flight/resources/resources.dart';
-import 'package:over_flight/router/app_hero_tags.dart';
-import 'package:over_flight/router/app_router.dart';
-import 'package:over_flight/router/app_routes.dart';
+import 'package:over_flight/router/router.dart';
 import 'package:over_flight/ui/widgets/buttons/filled_button_widget.dart';
+import 'package:over_flight/ui/widgets/buttons/google_button.dart';
 import 'package:over_flight/ui/widgets/text_fields/input_form_widget.dart';
+import 'package:over_flight/ui/widgets/title_section_widget.dart';
 
 class SignInFormWidget extends StatefulWidget {
   const SignInFormWidget({super.key});
@@ -22,8 +20,6 @@ class _SignInFormWidgetState extends State<SignInFormWidget> {
   @override
   Widget build(BuildContext context) {
     const divider = SizedBox(height: 28);
-    final theme = Theme.of(context);
-    final textStyle = theme.textTheme;
 
     return CustomScrollView(
       slivers: [
@@ -35,19 +31,9 @@ class _SignInFormWidgetState extends State<SignInFormWidget> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  ListTile(
-                    contentPadding: EdgeInsets.zero,
-                    title: Text(
-                      'Sign up for an account',
-                      style: textStyle.titleLarge,
-                    ),
-                    subtitle: Padding(
-                      padding: const EdgeInsets.only(top: 12),
-                      child: Text(
-                        'Hi, Enjoy your trip & easy book!',
-                        style: textStyle.titleMedium,
-                      ),
-                    ),
+                  const TitleSectionWidget(
+                    title: 'Sign up for an account',
+                    subtitle: 'Hi, Enjoy your trip & easy book!',
                   ),
                   divider,
                   InputFormWidget(
@@ -76,46 +62,16 @@ class _SignInFormWidgetState extends State<SignInFormWidget> {
                     title: "Password",
                     placeholder: "Enter your password",
                   ),
-                  TextButton(
-                    onPressed: () {},
-                    style: TextButton.styleFrom(
-                      overlayColor: Colors.transparent,
-                      padding: EdgeInsets.zero,
-                      side: BorderSide.none,
-                      shape: const RoundedRectangleBorder(
-                        borderRadius: BorderRadius.zero,
-                      ),
-                    ),
-                    child: const Text('Forgot password?'),
-                  ),
+                  divider,
                   divider,
                   FilledButtonWidget(
-                    onPressed: () =>
-                        AppRouter.go(context, AppRoutes.phoneVerification),
+                    onPressed: () {
+                      AppRouter.go(context, AppRoutes.phoneVerification);
+                    },
                     label: "Продолжить",
                   ),
                   const SizedBox(height: 20),
-                  SizedBox(
-                    width: double.infinity,
-                    child: TextButton.icon(
-                      onPressed: () {},
-                      style: TextButton.styleFrom(
-                        padding: const EdgeInsets.symmetric(vertical: 17),
-                        alignment: Alignment.center,
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(10),
-                          side: const BorderSide(
-                            color: Color(0xffEAEFF3),
-                          ),
-                        ),
-                      ),
-                      icon: SvgPicture.asset(AppIcons.google),
-                      label: Text(
-                        'Sign Up with Google',
-                        style: textStyle.titleMedium,
-                      ),
-                    ),
-                  ),
+                  const GoogleButton(),
                 ],
               ),
             ),
