+++ lib/ui/pages/phone_verification/widgets/keyboard/keyboard_data.dart
@@ -1,0 +1,11 @@
+enum KeyBoardButtonType { number, backspace, none }
+
+class KeyBoardButtonData {
+  final KeyBoardButtonType type;
+  final String? value;
+
+  KeyBoardButtonData({
+    required this.type,
+    required this.value,
+  });
+}
